# performance: Speed up Method Completions By Taking Advantage of Orphan Rules