@davidbarsky 2024-02-19
Right, okay: I addressed the review feedback, except the big one from Waffle. I'm pretty sure I've got a handle on checking fundamental types (I'll pilfer from it from the orphan implementation check), but I'm struggling to get a test with fundamental types working (see the changes in `crates/ide-completion/src/tests/flyimport.rs`). The expect side of things isn't the correct correct, but I can't seem to get the fixtures to be aware of a `#[fundamental]` `Box`—any tips?