@Veykril 2024-02-19
Aren't you missing the fundamental check?