@Veykril 2024-02-20
Oh wait, your test fixture is not using the fundamental type in any impl, so how should it show anything?