@davidbarsky 2024-02-20
Maybe a better of phrasing my question is "do I need to do anything special to introduce a new fundamental type in a fixture, or is the `#[fundamental]` annotation sufficient". Looking at some of the other tests, I'm inclined to say no.