@Veykril 2024-02-20
Just using the attribute is enough yes