@davidbarsky 2024-02-20
Alright, I updated the PR to include tests for `impl ForeignTrait for Box<LocalStruct>`, `impl ForeignTrait for &LocalStruct`, and `impl Into<A> for ()`. The first two tests pass without issue (in that the filtering already considers fundamental types), but the third test seemingly _does not_ work, despite `.into()` showing up in completions when I build a rust-analyzer binary with these exact changes. 

Notably, I _didn't_ need to make additional changes when filtering to consider fundamental types.